Move MyUserController and UserProfile files to v1 directory
