+++ config/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use app\models\User;
+use yii\rest\UrlRule;
 
 $params = require __DIR__ . '/params.php';
 $db = require __DIR__ . '/db.php';
@@ -61,6 +62,7 @@ $config = [
             'rules' => [
                 ['class' => 'yii\rest\UrlRule', 'controller' => 'v1/setup/country'],
                 ['class' => 'yii\rest\UrlRule', 'controller' => 'v1/setup/department'],
+                ['class' => 'yii\rest\UrlRule', 'controller' => 'v1/setup/my-user'],
             ],
         ],
     ],
