+++ controllers/UserController.php
@@ -5,7 +5,6 @@ namespace app\controllers;
 use app\helpers\MailHelper;
 use app\models\LoginForm;
 use app\models\User;
-use app\models\UserProfile;
 use app\rest\Controller;
 use Yii;
 use yii\base\Exception;
@@ -100,49 +99,5 @@ class UserController extends Controller
         if (!$user->save()) {
             return $this->validationError(Yii::t('app', 'Unable to save password'));
         }
-    }
-
-    /**
-     *
-     * @return UserProfile|array
-     * @author Levan Gogoladze <levanma98@gmail.com>
-     */
-    public function actionGetProfile($id)
-    {
-        $model = new UserProfile();
-        $profile = $model->getProfile($id);
-        if (!$profile) {
-            return $this->validationError(Yii::t('app', 'Unable to find user profile with this id'));
-        }
-        return $profile;
-    }
-
-    /**
-     *
-     * @author Levan Gogoladze <levanma98@gmail.com>
-     */
-    public function actionUpdateProfile() {
-        $request = Yii::$app->request->post();
-
-//        $user = Yii::$app->user->identity;
-//        return $user;
-        $user = User::findOne($request['id']);
-        $userProfile = $user->userProfile;
-        $user->email = $request['email'];
-
-        if($request['password'] and $request['password'] != '') {
-            $user->password = Yii::$app->security->generatePasswordHash($request['password']);
-        }
-
-        $userProfile->firstname = $request['firstName'];
-        $userProfile->lastName = $request['lastName'];
-        $userProfile->phone = $request['phone'];
-        $userProfile->mobile = $request['mobile'];
-        $userProfile->birthday = $request['birthday'];
-        $userProfile->aboutMe = $request['aboutMe'];
-        $userProfile->hobbies = $request['hobbies'];
-
-
-        return Yii::$app->user->identity->email;
     }
 }
